Split up nexus components into separate files.
